Add style for signup form, small refactoring of auth form
