+++ shop-frontend/src/Components/Auth/Auth.tsx
@@ -1,6 +1,7 @@
 import React, { useState } from 'react'
 import './Auth.css'
 import { SignupForm } from './SignupForm/SignupForm'
+import { AuthForm } from './AuthForm/AuthForm'
 
 export const Auth = () => {
  
@@ -40,36 +41,5 @@ export const Auth = () => {
   )
 }
 
-export const AuthForm = () => {
-    
-    const [inputLogin, setInputLogin] = useState('')
-    const [inputPassword, setInputPassword] = useState('')
-    
-    return (
-        <form>
-        <div className="form_auth__inputs">
-          <div className="form_auth__input__login">
-            <label className="form_auth__label login">Имя</label>
-            <input
-              className="input_login"
-              type="text"
-              value={inputLogin}
-              onChange={(e) => setInputLogin(e.currentTarget.value)}
-            />
-          </div>
-          <div className="form_auth__input__password">
-            <label className="form_auth__label password">Пароль</label>
-            <input
-              className="input_password"
-              type="text"
-              value={inputPassword}
-              onChange={(e) => setInputPassword(e.currentTarget.value)}
-            />
-          </div>
-        </div>
-        <button className="form_auth__button_action">Войти</button>
-      </form>
-    )
-  
-}
+
