+++ shop-frontend/src/Components/Auth/AuthForm/AuthForm.css
@@ -1,0 +1,38 @@
+.form_auth__inputs {
+    min-width: 45%;
+    display: flex;
+    justify-content: space-between;
+    flex-direction: column;
+    margin-top: 50px;
+}
+
+.form_auth__input__login, 
+.form_auth__input__password {
+    display: flex;
+    flex-direction: column;
+    height: 60px;
+    margin-bottom: 20px;
+}
+
+.form_auth__label {
+    opacity: 0.7;
+}
+
+.input_login, 
+.input_password {
+    outline: 1px solid var(--color-font);
+    opacity: 0.8;
+    height: 40px;
+    width: 350px;
+}
+
+.form_auth__button_action {
+    border-radius: 7px;
+    text-align: center;
+    margin-top: 50px;
+    background-color: var(--color-tan);
+    width: 195px;
+    height: 50px;
+    text-transform: uppercase;
+    letter-spacing: 3px;
+}
